requirements.txt
jupyter==1.0.0
numpy==1.24.3
pandas==2.0.3
matplotlib==3.7.1
seaborn==0.12.2
openai==1.3.0
python-dotenv==1.0.0
openpyxl==3.1.5
tqdm==4.66.1
requests==2.31.0
aiohttp==3.9.1
boto3==1.34.0
pymupdf==1.23.0 